+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased: mitmproxy_rs next
 
+- Various fixes for Protobuf number encoding.
 
 ## 29 April 2025: mitmproxy_rs 0.12.2